pyproject.toml
[project]
name = "pdf-editor"
version = "0.1.0"
description = "Self-hosted browser PDF editor monorepo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "pypdf>=5.1.0,<6.0.0",
  "PyMuPDF>=1.24.10,<2.0.0",
  "Pillow>=11.0.0,<12.0.0",
  "python-multipart>=0.0.18,<1.0.0",
]

[tool.uv]
dev-dependencies = []
